Split virtual-mem-related code into its own files.
